+++ MutaroApp/MutaroUITests/MutaroUITestsLaunchTests.swift
@@ -1,0 +1,32 @@
+//
+//  MutaroUITestsLaunchTests.swift
+//  MutaroUITests
+//
+//  Created by minguk-kim on 2023/01/09.
+//
+
+import XCTest
+
+final class MutaroUITestsLaunchTests: XCTestCase {
+
+    override class var runsForEachTargetApplicationUIConfiguration: Bool {
+        true
+    }
+
+    override func setUpWithError() throws {
+        continueAfterFailure = false
+    }
+
+    func testLaunch() throws {
+        let app = XCUIApplication()
+        app.launch()
+
+        // Insert steps here to perform after app launch but before taking a screenshot,
+        // such as logging into a test account or navigating somewhere in the app
+
+        let attachment = XCTAttachment(screenshot: app.screenshot())
+        attachment.name = "Launch Screen"
+        attachment.lifetime = .keepAlways
+        add(attachment)
+    }
+}
